rewrite our custom store with get/set, because update(null) resets the store
contrary to popular belief, returning null/undefined from update(s => {}) will
just set that store to null, (i.e. you have no way to cancel the update), so we
need to do a whole get/set dance to get the same result… thank you svelte.

this was found by @blackm2021 that actually looked into the store code and
found it to be quite buggy, more on this later.