CSS and style changes on MediaForm, MediumVideo and Media modules
The general idea is to have a form to submit new information related to the media objects we use on the analysis.
In order to do that, we've modified the style so we give prominence to the media files and to make sure the form only takes up spaces when is strictly needed.

I've commented the SvgOverlay component because it seems we need to update its dimension setups to avoid breaking the parent component's CSS.